requirements.txt
anthropic==0.15.0
MetaTrader5==5.0.45
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.7.3
seaborn==0.13.0
scipy==1.11.3
statsmodels==0.14.0
pytz==2023.3
TA-Lib==0.4.26